move ds schemas together with resource one
